Implemented plugin configuration in fennel
