+++ stow/nvim/.config/nvim/fnl/dotfiles/init.fnl
@@ -2,6 +2,24 @@
     {require [aljendro.config.functions
               dotfiles.core.settings
               dotfiles.core.keybindings
+              dotfiles.plugins.surround
+              dotfiles.plugins.ultisnips
+              dotfiles.plugins.cmp
+              dotfiles.plugins.fugitive
+              dotfiles.plugins.gitsigns
+              dotfiles.plugins.harpoon
+              dotfiles.plugins.hop
+              dotfiles.plugins.lspconfig
+              dotfiles.plugins.lualine
+              dotfiles.plugins.neoscroll
+              dotfiles.plugins.nrrwrgn
+              dotfiles.plugins.nvim-tree
+              dotfiles.plugins.surround
+              dotfiles.plugins.telescope
               dotfiles.plugins.tokyonight
-              aljendro.config.core]})
+              dotfiles.plugins.treesitter
+              dotfiles.plugins.ultisnips
+              dotfiles.plugins.vim-test
+              dotfiles.plugins.vim-visual-multi
+              dotfiles.plugins.vimux]})
