+++ stow/nvim/.config/nvim/fnl/dotfiles/plugins/lspconfig.fnl
@@ -190,5 +190,3 @@
                            :typescript [prettier eslint]
                            :typescriptreact [prettier eslint]}}})
 
-
-
