+++ stow/nvim/.config/nvim/fnl/dotfiles/plugins/neoscroll.fnl
@@ -46,3 +46,4 @@
     (set toggled (not toggled)))
 
 (ToggleSmoothScroll)
+
