+++ stow/nvim/.config/nvim/fnl/dotfiles/plugins/nrrwrgn.fnl
@@ -1,0 +1,8 @@
+(module dotfiles.plugins.nrrwrgn
+  {autoload {{: kmap} dotfiles.core.common}})
+
+(set vim.g.nrrw_rgn_vert true)
+(set vim.g.nrrw_rgn_resize_window "relative")
+
+(kmap :x "<leader>z" "<Plug>NrrwrgnDo" {:silent true})
+
