+++ stow/nvim/.config/nvim/fnl/dotfiles/plugins/surround.fnl
@@ -1,0 +1,9 @@
+(module dotfiles.plugins.surround
+  {autoload {{: kmap} dotfiles.core.common}})
+
+(kmap :n "<leader>'" "ysiw'" {:silent true})
+(kmap :n "<leader>\"" "ysiw\"" {:silent true})
+(kmap :n "<leader>`" "ysiw`" {:silent true})
+(kmap :n "<leader>)" "ysiw)" {:silent true})
+(kmap :n "<leader>}" "ysiw}" {:silent true})
+
