+++ stow/nvim/.config/nvim/fnl/dotfiles/plugins/treesitter.fnl
@@ -11,4 +11,3 @@
                                      :node_decremental "<leader>u"
                                      :node_incremental "<leader>i"}}})
 
-
