+++ stow/nvim/.config/nvim/fnl/dotfiles/plugins/ultisnips.fnl
@@ -1,0 +1,6 @@
+(module dotfiles.plugins.ultisnips)
+
+(set vim.g.UltiSnipsRemoveSelectModeMappings false)
+(set vim.g.snippetPath (.. (os.getenv "HOME") .. "/.config/nvim/ultisnips"))
+(set vim.g.UltiSnipsSnippetDirectories [vim.g.snippetPath])
+
