+++ stow/nvim/.config/nvim/fnl/dotfiles/plugins/vim-visual-multi.fnl
@@ -1,0 +1,10 @@
+(module dotfiles.plugins.vim-visual-multi)
+
+;; Settings
+(set vim.g.VM_theme "neon")
+(set vim.g.VM_silent_exit true)
+
+;; Keybindings
+(set vim.g.VM_leader "<bslash>")
+(set vim.g.VM_mouse_mappings true)
+
